+++ src/lotto/view/InputReader.js
@@ -1,7 +1,7 @@
 import { input } from '../../util/IOUtil.js';
 import { INPUT_MESSAGE } from '../constants/Message.js';
 import CommonValidator from '../validate/CommonValidator.js';
-import { retryIfOccurredError } from '../RetryHelper.js';
+import { retryIfOccurredError } from '../helper/RetryHelper.js';
 import LottoWinningNumbers from '../model/LottoWinningNumbers.js';
 import LottoPayment from '../model/LottoPayment.js';
